More Pantry updates for Region replacement (#1269)
Add some status endpoints so that Nexus can ask the Pantry some
questions:

- how many volumes and job handles are you currently holding on to?

- for a given volume, is it active, and how many job handles refer to
it?

Also:

- add an `attach_activate_background` endpoint to the Pantry, so
activation can occur in a job instead of blocking the activate endpoint.
this will be used by Nexus when it knows that a volume needs
reconcilation.

- add a `replace` endoint to the Pantry, so that Nexus can call
`target_replace` for volumes attached to the antry

Additionally, move `ReplacementResult` to crucible-client-types, and
have `target_replace` return that instead of (). This will be returned
by Propolis so that Nexus can tell what the status of a requested
replacement is.

The Pantry also now keeps track of if it has seen a Volume active or
not. It also does not `flush` before calling `deactivate` anymore, as 
`deactivate` already submits a final flush.