Remove `ReadResponse` in favor of `RawReadResponse` (#1212)
In #1206, we reorganized messages to put bulk data in a single place. In
other words, a read response is no longer split into a bunch of
individual block-size buffers; it's all in a single contiguous
allocation.

However, to minimize the size of the PR, we kept the existing `struct
ReadResponse`, tweaking it to borrow from the contiguous allocation.

This PR removes the vestigal `ReadResponse` in favor of using the
`RawReadResponse` everywhere. This is a net LOC decrease, and means that
we can copy into our guest buffer in a single `memcpy` (ownership
permitting), instead of multiple block-sized copies.

The majority of changes (by LOC) are calls to `process_ds_completion`,
which previous took a `Vec<ReadResponse>` and now takes a
`RawReadResponse`.

The part of the change requiring close review is the [new
`Buffer::write_read_response`](https://github.com/mkeeter/crucible/blob/19926c38575e6b43f976ee9be66ed252ac39a14c/upstairs/src/lib.rs#L1587),
which takes care to avoid `memcpy` when possible.