Refactor downstairs reads (#1283)
This is analogous to #1272 , but for reads instead of writes.

By the time we send a read to an extent, we want it to be contiguous and
tightly packed; this makes extent code simpler and lets us do bulk
reads.

The new data flow is `Message::ReadRequest` → `RegionReadRequest` →
`ExtentReadRequest` → `ExtentReadResponse` → `Message::ReadResponse`.
Allocation is minimized: we build a single `BytesMut` then use `split()`
and `unsplit()` to pass it to each extent.

More generally, these two PRs are a stepping stone towards non-blocking
Downstairs IO, i.e. having IOs to multiple extents pending
simultaneously.