Refactor downstairs writes (#1272)
crucible#1206 reorganized the `Message::Write` variant so that the bulk data was
all packed at the end, and could be sent in a single write (instead of incurring
the cost of serialization).

The `Downstairs` then took that data and split it into a
`Vec<crucible_protocol::Write>` objects, which each look like this:
```rust
pub struct Write {
    pub eid: u64,
    pub offset: Block,
    pub data: Bytes,

    pub block_context: BlockContext,
}
```

Note that this `Write` does not do an extra copy; the `data: Bytes` is
borrowed from the received data (this is good).

However, it's still an awkward API: a single contiguous `Bytes` object
arrives on the wire, then we split it into a bunch of smaller
(block-sized) chunks, then we write those chunks to contiguous blocks
with `pwritev`.

Put a different way, it's awkward to have `ExtentInner::write` take a
`&[crucible_protocol::Write]`: the extent implementation has to jump
through hoops to reassemble contiguous writes from the incoming slice of
writes. As a smaller papercut, it's also awkward that this struct
contains an `eid`: what should happen if someone calls
`ExtentInner::write` with an `eid` that doesn't match?

This PR reorganizes our writes into two new types in the
`crucible-downstairs` crate:
```rust
/// Deserialized write operation, contiguous to a single extent
///
/// `data` should be a borrowed section of the received `Message::Write`, to
/// reduce memory copies.
#[derive(Debug, Clone, Eq, PartialEq)]
pub(crate) struct ExtentWrite {
    pub offset: Block, // starting position
    pub block_contexts: Vec<BlockContext>, // one per block
    pub data: bytes::Bytes, // num blocks * block size
}

/// `ExtentWrite` plus an extent tag
#[derive(Debug, Clone, Eq, PartialEq)]
pub(crate) struct Write {
    extent: u64,
    write: ExtentWrite,
}
```

The `ExtentWrite` is represents a set of contiguous writes, because it
simply doesn't store one offset per block; it just has a starting
offset. Note that it does not contain an extent tag, so
`ExtentInner::write` no longer needs to think about what happens if
there's a mismatch. The target extent lives in a new
`crucible_downstairs::Write` object, where it's used for dispatch at the
`Region` level (but not passed to the `Extent`).

The process of splitting the write into contiguous, separated-by-extent
subwrites still has to happen, but it's now a conversion from
`Message::Write → Vec<Write>` that happens when messages are received.

This means that the extent implementations can use `pwrite` instead of
`pwritev`, and no longer need to consider whether the incoming writes
are contiguous.

There should be no change to on-disk formats; this is purely a change to
how we wrangle data from the wire into the extent data file.