Update crutest replace test, and mismatch printing.
Updated the replace test to better reflect replacing a downstairs
when IO is underway.  We spawn a task to do the replacement and
wait for it to finish while the main task just sends IO in a
loop.

Updated the crutest block mismatch printing code to detect a
mismatch range and just print a summary at the end of the range
instead of printing a line for every byte that does not match what is
expected.

Updated the CI tests for live repair to make a little larger region
so we can catch IO in progress while doing live repair.
Updated the test to also collect dtrace stats.