Allow replacement of a target in a read_only_parent (#1281)
Update the VCR validation and replacement code to accept changes to the
read_only_parent field.

For this, I was able to re-use the same code that was checking
sub_volumes, by pulling that common part out to a sub-function which we 
could use to also check read only parents. The meat of the comparison tests
is all the same, just moved down. Since we now compare both sub_volume
and read_only_parent, we have to then compare the results with each other
and be sure that only one of the two changes.

New tests added for read only parent replacement.

Cleaned up a few copy paste errors in other test comments.
Simplified a bunch of VolumeConstructionRequest assignments.
---------

Co-authored-by: Alan Hanson <alan@oxide.computer>