Move disconnect logic to the Downstairs (#1332)
This PR moves the responsibility for "disconnecting when an error
occurs" into the `struct Downstairs`.

Previously, `Downstairs::on_message_for` returned a `Result`. If the
result was an error (or `false`), the IO task would exit, but the
Downstairs wouldn't necessarily notice until someone tried to
communicate with that IO task.

This PR makes `Downstairs::on_message_for` a fire-and-forget operation
from the perspective of the IO task, returning nothing. If the message
_does_ result in an error, the `Downstairs` itself will note the error
and drop that connection. When the connection is dropped, the
cancellation token held in the `ConnectionState` will automatically stop
the IO tasks.