Support arbitrary Volumes during replace compare
This commit changes `compare_vcr_for_update` to (attempt) to support
arbitrary Volume hierarchies. A bunch of tests were added to ensure that
the cases that Nexus creates are covered, but I'm not 100% sure that all
arbitrary Volumes are covered.

This commit also adds the intentional behaviour of ignoring the
generation number of the read-only part of a Volume during comparison
for replacement.

Fixes #1342