Move Work into ConnectionState (#1340)
This moves the `struct Work` into the `struct ActiveConnection`.

It was previously in an `ActiveDownstairs`, which was stored in
`active_upstairs: HashMap<Uuid, ActiveUpstairs>`.

Putting it into the `ActiveConnection` consolidates where per-connection
data lives, since we don't actually want to reuse `Work` if someone
reconnects with a matching `Uuid`. This is a step towards having the
`ActiveConnection` do most of the per-connection work; for example, this
PR moves `is_message_valid` and `add_work` into `ActiveConnection`.

Because we only have one task accessing the `Downstairs`, we can then
remove a bunch of checks against being interrupted mid-operation: we
have a `&mut Downstairs`, so no one is interrupting us. In other words,
many functions become infallible once we've checked that we're operating
on an active connection.