Remove `DownstairsWork` from map when handling it (#1361)
Previously, we borrowed the `DownstairsWork` from the map when doing IO
work.

In this PR, we switch to a strongly-owned model: the `DownstairsWork` is
removed from the map when it's ready to be handled. Removing the job
from the map also eliminates a `clone()` in `get_ready_job`, which could
be expensive for large writes (where the `Vec<BlockContext>` is
non-trivial). To prevent clones from sneaking in, we remove
`#[derive(Clone)]` from `IOop` and `DownstairsWork`.

This is a building block for processing ready IOs immediately as they
arrive, instead of putting them in the map at all.