Replace a downstairs before we are active.
Allow a downstairs to be replaced if the request comes in before the
upstairs is active.  If reconciliation is in progress, it will be aborted
and then restarted.

Created two new crutest tests that will test replacement before active
for both before reconciliation has started and after reconciliation has
started.

Updated the crutest cli to allow us to send an activation without waiting
for the result (which then allows us to trigger a replacement).

Added a new test, test_replace_special.sh that will spin up downstairs
and then call crutest and test the replacement of a downstairs while
reconciliation is underway.

Added two tests to `test_up.sh`. One that will test replacement of a
downstairs before activation and before reconciliation. A second test will
test replacement before activation but during reconciliation.

Added the test_replace_special.sh to the nightly test.