Add `BlockOffset` and `BlockIndex` types (#1379)
The `Block` type is very overloaded:

- Sometimes, its value represents an offset relative to the start of an
  extent
- Other times, its value represents an absolute block index
- In some cases, it represents an extent size!
- `Block` contains both a `value` and a `shift`; `shift` usually means
  _log2(block size)_
  - `shift` is rarely used and only sometimes checked
  - It's basically only relevant when `Block` represents an extent size
- However, it's always present, so we have to provide it to `Block::new`
  even when irrelevant

This PR adds two new types to clarify the meaning of `Block`:

- `BlockIndex` is an absolute block index
- `BlockOffset` is an offset relative to the start of an extent

Both of these types are wrappers around a single `u64` and **do not**
store block size. This is good, because it simplifies the conceptual
space: since the `Block` argument includes the block size, we have to
handle the case where the block size is **wrong**. If we simply don't
store the block size, then we eliminate all of that logic.

The rest of this PR is using `BlockIndex` and `BlockOffset` types
wherever possible, and removing superfluous functions on the original
`Block` type.

WARNING: This is a change to wire format, because we're using
`BlockOffset` (one `u64`) instead of `Block` (a `u64` + `u32`) in
various messages.

This does **not** change the `BlockIO` API, although I'm planning to do
that in a subsequent PR.

See also https://github.com/oxidecomputer/crucible/issues/923 , although
I don't change `RegionDefinition` in this PR (because the existing
format has been serialized to disk).