Replace a downstairs before we are active. (#1381)
Allow a downstairs to be replaced if the request comes in before the
upstairs is active.
Fix for: https://github.com/oxidecomputer/crucible/issues/871
(Also fixed: https://github.com/oxidecomputer/crucible/issues/1375)
(Also fixed: https://github.com/oxidecomputer/crucible/issues/1376)

This actually was a pretty small change to the upstairs code itself as
the general idea of replacement before active fits within the overall
"make all downstairs consistent before starting".

If the activate has not been requested from the guest, then we can do
the replacement in place (as we have not started negotiation yet). If
reconciliation is in progress, it will be aborted and then restarted. The
code already checked for any state change in a downstairs and will
abort a reconciliation in progress. All we had to change here was
allowing the reconciliation to restart after a replacement.

Created two new crutest tests that will test replacement before active
for both before reconciliation has started and after reconciliation has
started. Added a clone of the logger for crutest itself to use.

Updated the crutest cli with a new 'activate-request` command that
will allow us to send an activation without waiting for the result. This
then allows us to trigger a replacement from the cli before an upstairs
is active.

Added a new test, test_replace_special.sh that will spin up downstairs
and then call crutest and test the replacement of a downstairs while
reconciliation is underway.

Added two tests to `test_up.sh`. One that will test replacement of a
downstairs before activation and before reconciliation. A second test
will test replacement before activation but during reconciliation.

Added the test_replace_special.sh to the nightly test.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>