Update `BlockIO` trait to use `BlockIndex` type (#1385)
This PR updates `trait BlockIO` to use `BlockIndex` arguments (instead
of `Block`).

`BlockIndex` was introduced in #1379 and has the specific meaning of "an
absolute block index". Unlike `Block`, it does not carry a size, because
it's weird for all of our operations to have the semantics of

> Do a write at this absolute block, and also here's the block size, so
maybe you should check to make sure it's still valid?

By passing _only_ an absolute block index, we obviate any questions
about checking and handling the `Block::size` member.

/!\ This will require minor changes in Propolis, which I can do once this
is merged.