Remember that a replaced downstairs is replaced
Leave a downstairs in state Replaced after a replacement request
until we connect to the new downstairs and either validate it is
correct or determine it is bad.

This fixes a bug where a replaced downstairs would go to failed,
then cause problems when we eventually connect to that downstairs.