Remove `Replay` state (#1430)
We never actually dwell in the `Replay` state; we enter `Replay` then
immediately enter `Active` one line later.

The actual work of replaying jobs is done elsewhere, in
`Downstairs::reinitialize`. As such, the `Replay` state doesn't need to
exist, and is kinda misleading; for example, we can't actually
transition from `Replay` to `Faulted` or `Deactivated`, despite the flow
diagram.