Make backpressure ignore `Offline` downstairs (#1431)
One of my complaints in #1258 is that the `Offline` → `Faulted`
transition depends on reaching a certain number of jobs in the queue,
but accumulating those jobs is subject to backpressure.

This has a few awkward consequences:
- The time it takes to transition from `Offline` → `Faulted` depends on
backpressure tuning, and we have to pick backpressure parameters that
make this take a reasonable amount of time (tested in
`check_offline_timeout`)
- The offline Downstairs slows down the other Downstairs based on the
number of jobs in its queue, meaning the system as a whole gets slower
and slower until that Downstairs is marked as faulted

This PR tweaks the backpressure implementation so that only `Active`
Downstairs are counted. To do so, we move the `write_bytes_outstanding`
to a per-client field. The backpressure implementation is also moved to
a separate module, since we'll be adding more to it in the future.

The `check_offline_timeout` test is no longer relevant and has been
removed.