Make crutest use BlockIO trait instead of a Geust
Update crutest to use BlockIO instead of Guest.  This involved adding a
few more methods to the BlockIO trait, moving a few guest specific methods
over to BlockIO, and adding support for these new methods in places
where they did not have them.

This should result in no differences at run time as we are not doing
anything different than before.  This just enables future work where
we can use a Volume in crutest instead of a Guest.