Address PR comments, new activate behavior
Address PR comments

We now allow an activate command to succeed if the upstairs is already
active.  In the past this returned Err(UpstairsAlreadyActive), but
now it just returns Ok(()).  From the volume layer the behavior is the
same as the Volume layer would only attempt to activate if the sub-volume
was not yet active, now we have just formalized this behavior into the
activate() call itself.

Changed the return error for an active request to return
UpstairsActivateInProgress if there is already an activation in progress.
We store the BlockRes to respond to an activation request until an upstairs
has activated and respond over that channel when we get to active, if the
caller has chosen to not wait on that channel and instead try to activate
again, we will return activation in progress (as we already have a BlockRes
we use when the activation is complete).

Removed all the conditional_activate() support.

Changed the behavior of BlockOp::GoActiveWithGen, which is only used
in tests.  We now allow this activation if the upstairs is not currently
active (and update it to use the new gen).  If the upstairs is already
active with the same gen, then that is also okay.  If the upstairs is
active with a different gen, then we fail the BlockOp::GoActiveWithGen.