Remove `fn in_progress` to simplify logic (2/n) (#1463)
(staged on top of #1460)

Previously, we built a `DownstairsIO` relatively early in the process of
submitting / sending a job. There's a bunch of boilerplate involved,
which is usually the same for all IO types.

This PR reorganizes code to keep using `IOops` until the very last
minute, _then_ builds the `DownstairsIO` at the end of `enqueue` and
`enqueue_repair`. It simplifies code and removes duplication, but
shouldn't be a functional change.