Update more tests to use dsc (#1480)
Give crutest the list of targets it will be operating on so some tests
that need that list of targets can still get it.

Updated more tests to use dsc to determine targets instead of having to
provide them on the command line.

Updated most of test_up.sh to use dsc for targets instead of supplying
them on the command line.  There is still some work required to get
hammer tests to use dsc, if we decide to do that.