Remove `JobId` from `DownstairsIO` (1/2) (#1481)
`DownstairsIO` are stored in a `BTreeMap<JobId, DownstairsIO>`, so
storing the `JobId` in the `DownstairsIO` is superfluous.

This PR removes the `ds_id` field. Most of the LOC changes are tweaking
unit tests, which previously pulled the active jobs array and looked at
the `ds_id` field; instead, the various job submission functions return
a `JobId` that the tests store and use.

In addition, I added a `get_deps(ds_id: JobId)` helper function to
reduce boilerplate.