Remove `GuestWorkId` (2/2) (#1482)
(staged on top of #1481)

`GuestWorkId` is always 1:1 with `JobId`, so let's get rid of it!

It's not exposed to the guest or used by anything outside of Crucible
internals.

The only exception is DTrace probes, which now use the `JobId`
everywhere (instead of a mix of `JobId` and `GuestWorkId`). The
`gw__*__start` probes now fire immediately after a `JobId` is assigned,
rather than when the `GuestWorkId` was assigned.