Remove remaining `Arc<Volume>` instances (#1500)
Nested volumes use an `Arc<dyn BlockIO + Send + Sync>` to represent the
inner `BlockIO` device.

Previously, we implemented the `BlockIO` trait on the `struct Volume`
itself, so we needed to make an `Arc<Volume>` to create nested volumes.

This PR moves the `BlockIO` implementation to the `VolumeInner`. Because
the `VolumeInner` is already stored in an `Arc`, we can now cast from
`Arc<VolumeInner>` to `Arc<dyn BlockIO + Send + Sync>`, removing the
extra level of indirection.