Simplify `CompleteJobs` (#1493)
We never actually use the "jobs before `last_flush` count as completed"
property of `CompleteJobs`.

From discussions in chat, this may have been a legacy from when we sent
literally every previous job in the dependency list.
Now that flushes act as a full dependency barrier, it shouldn't be
possible to depend on anything before a flush.

This PR removes special-casing of `last_flush`; instead it becomes first
among equals in `completed: Vec<JobId>`. Most of the changes are to edit
test cases that violate the "flushes are a full barrier" property.