Add region set options to more tests. (#1496)
Give test_live_repair, test_replace_special, and test_replay an option
to indicate how many region sets should be created for the test. This
allows us to use the same script to test single and multiple
sub-volumes.

Updated a few of the tests to no longer run a smaller test in a loop,
but instead just run a longer test. This makes better sense when testing
many sub-volumes as a smaller test might not cover everything. It has
the side effect of not printing regular status messages, but if we find
that is important we can come up with another way of displaying test
progress.

Future PRs will use these options to increase test coverage on multiple
sub-volumes.

---------

Co-authored-by: Alan Hanson <alan@oxide.computer>