Add `IOop::Barrier` (#1494)
The `Barrier` op does no work, but serves as a full dependency barrier
(like a flush) so that the Downstairs can reset its list of completed
jobs. Note that it does _not_ update `last_flush` or retire jobs in the
Upstairs, because it doesn't actually persist data to disk.

This is a building block for removing automatic flushes (see RFD 518).