Split out Downstairs-specific stats (#1511)
This removes another bit of awkward plumbing!

Previously, we had to pass an `&UpStatOuter` to any functions in the
`Downstairs` that needed to update our Oximeter statistics. This PR
splits the `UpStatOuter` into upstairs versus downstairs components,
then stores a (shared) handle to the downstairs half in the `Downstairs`
itself.

The Oximeter interface remains unchanged.