Add unit test for VersionMismatch (#1524)
While looking at negotiation, I noticed that `VersionMismatch` is
load-bearing: the Downstairs will send it if it receives a `HereIAm`
with an incorrect version. As such, we should be unit-testing that its
representation doesn't change (same as `HereIAm` and `YesItsMe`). I also
made the `YesItsMe` test more exhaustive, testing the full encoding of
both IPv4 and IPv6 addresses.