Target replacement support for VCRs with multiple sub_volumes
Updated the volume layer code to support replacement of a target in
a VCR that has multiple sub_volumes in it.  Before this we only
supported a single sub_volume.

Most of the changes here are for tests.  I've added a loop around vcr
replacement tests to create multiple sub volumes and iterate through
the test for different VCR configurations.

In addition, it is now required that the generation number change for
all sub-volumes when any target is changed.